pyproject.toml
[tool.poetry]
name = "authsignal"
version = "4.0.3"
description = "Authsignal Python SDK for Passwordless Step Up Authentication"
authors = ["justinsoong <justinsoong@gmail.com>"]
license = "MIT"
readme = "readme.md"
documentation = "https://docs.authsignal.com/"
homepage = "https://www.authsignal.com/"
repository = "https://github.com/authsignal/authsignal-python"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
pyjwt = "^2.8.0"
pyhumps = "^3.8.0"

[tool.poetry.dev-dependencies]
responses = "^0.24.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
